added folder addition and deletion check

+++ src/components/FolderComponent.tsx
@@ -1,17 +1,25 @@
 import DeleteIcon from "@mui/icons-material/Delete";
 import AddIcon from "@mui/icons-material/Add";
 import EditIcon from "@mui/icons-material/Edit";
-import { deleteFolder, updateFolder } from "../utils/api";
+import { createFolder, deleteFolder, updateFolder } from "../utils/api";
 import { useState, KeyboardEvent } from "react";
 import { Input } from "@mui/material";
 import { useDispatch } from "react-redux";
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
-export default function FolderComponent({ folder }: any) {
+export default function FolderComponent({ folder, hasChildren }: any) {
   const [editing, setEditing] = useState(false);
   const dispatch = useDispatch();
 
-  const handleDelete = async () => {
+  const handleDelete = async (e: any) => {
+    if (editing) return;
+    e.stopPropagation();
+
+    if (hasChildren) {
+      alert("You can't delete a folder with children!");
+      return;
+    }
+
     await deleteFolder(folder);
     dispatch({ type: "DELETE_FOLDER", payload: folder });
   };
@@ -38,6 +46,20 @@ export default function FolderComponent({ folder }: any) {
     }
   };
 
+  const handleAdd = async (e: any) => {
+    // e.stopPropagation();
+    const newFolder = {
+      name: "New Folder",
+      id: Math.trunc(Math.random() * 100000).toString(),
+      isFolder: true,
+      parentId: folder.id,
+    };
+
+    console.log("newFolder", newFolder);
+    const addedFolder = await createFolder(newFolder);
+    dispatch({ type: "ADD_FOLDER", payload: addedFolder });
+  };
+
   return (
     <div className="FolderComponent">
       {editing ? (
@@ -48,10 +70,10 @@ export default function FolderComponent({ folder }: any) {
           onClick={(e) => e.stopPropagation()}
         />
       ) : (
-        folder.name
+        <span>{folder.name}</span>
       )}
 
-      <AddIcon fontSize="inherit" onClick={(e) => e.stopPropagation()} />
+      <AddIcon fontSize="inherit" onClick={handleAdd} />
       <EditIcon fontSize="inherit" onClick={handleEdit} />
       <DeleteIcon fontSize="inherit" onClick={handleDelete} />
     </div>
